packages/example_model/pyproject.toml
[project]
name = "example-model"
version = "0.1.0"
description = "Example branching process model"
readme = "README.md"
authors = [
    { name = "Eric Mooring", email = "pgv5@cdc.gov" },
    { name = "Beau B. Bruce", email = "lue7@cdc.gov" },
]
requires-python = ">=3.12"
dependencies = ["numpy>=2.3.4", "cfa-mrp", "calibrationtools"]

[project.scripts]
example_model = "example_model.example_model:main"

[tool.uv.sources]
calibrationtools = { workspace = true }

[build-system]
requires = ["uv_build>=0.9.21,<0.10.0"]
build-backend = "uv_build"
